Project/Backend/requirements.txt
Django>=2.2
<<<<<<< HEAD
djangorestframework
pika
django-cors-headers
=======
>>>>>>> origin/feature/data_generation
